ncc: first milestone
the compiler now successfully compiles itself using its own
preprocessor and standard library. ad hoc testing shows that
the compiler, when compiling itself with a version compiled
with itself, produces the same code as the compiler compiled
by gcc compiling itself- with only insignificant variations
caused by the order of evaluation of function arguments being
different between gcc and ncc. (this causes the assignment of
pseudo-registers in a slightly different order, leading to
slightly different register allocation decisions.) this gives
a high degree of confidence that the compiler is fundamentally
sound, though, of course, many bugs are doubtless still lurking.

i've reset the repository to begin at this point as the mode of
development will now change, and further commits will be subject
to regression testing: this is especially important as the next
round involves implementing several optimization passes.

Signed-off-by: Charles E. Youse <charles@gnuless.org>