cc1: amd64/gen.c: attempt to combine subraction in selxn cache
the AMD64 generator builds operands by interpreting adds and shift
and mutating them into effective addresses. we now add subtraction
to the mix, which covers a few additional cases we overlooked.

this exacerbates a problem where the code generator selects, e.g.,

            leaq -4(%rax), %rax

instead of the more obvious

            subq $4, %rax

when it sees different source and target registers in the IR that
are only later coalesced by the register allocator. this is easy
to fix in a planned peephole stage specific to AMD64.

Signed-off-by: Charles E. Youse <charles@gnuless.org>