cc1: AMD64: revert LEA instructions to simpler primitives
as explained in #7, the code generator makes suboptimal choices
when working with pseudo-registers which are later coalesced.
we add a quick pass after registers are allocated to revert the
use of effective address computations rather than their more
straightforward add/shift counterparts.

along the way, added a new struct target field which specifies
a function to run to optimize after register allocation is done.

fixes: #7

Signed-off-by: Charles E. Youse <charles@gnuless.org>