cc1: AMD64: do not rewrite MOVs as XORs if condition codes are live
mov $0, %r can't be rewritten as xor %r, %r if the condition codes
are live, as the latter will trash them. i doubt this situation will
arise in practice in the current implementation (and indeed, it does
not change any of the regression files) but this would come back to
bite us as the transformations get more complex.

Signed-off-by: Charles E. Youse <charles@gnuless.org>