cc1: insn.[ch]: insn_uses_cc() now returns ccset instead of bool
it now returns the set of condition_codes actually used,
rather than a simple boolean indicating whether any are
used or not. along the way, removed I_FLAG_USE_CC as it
is redundant; the only IR insns that use the CCs are the
I_SET_CCs, which are already flagged.

updated the AMD64 target function to conform, although it
only provides course granularity (all codes or no codes).

Signed-off-by: Charles E. Youse <charles@gnuless.org>