cc1: add I_TEST IR instruction, and add optimization to use it
both AMD64 and ARM32 (our primary targets) support the concept
of a bitwise AND instruction sets the condition codes but tosses
the result. we add I_TEST to the IR, and modify the AMD64 code
generator to support it.

we add a new target-independent optimization which rewrites ANDs
to TESTs as appropriate, as well as some supporting routines
which will be handy in other transformations.

fixes: #11

Signed-off-by: Charles E. Youse <charles@gnuless.org>