cc1: gen.c: emit I_MOVE instead of I_CAST for discrete downcasts
it's sufficient to make a copy of the value. we gain nothing
by marking it in the IR as anything other than an I_MOVE, as
that just prevents copy propagation and such from recognizing
that the destination is, in fact, a copy.

this is on net a win, though it exposes a minor flaw in the
coalescing logic which results in qword moves in a couple of
places where dword moves would be sufficient. will fix.

Signed-off-by: Charles E. Youse <charles@gnuless.org>