cc1: eliminate useless mutual copies arising from false edges
the register allocator sometimes backs itself into a corner with
constructs like (AMD64 code):

			movq %rdi, %i16q
			movq %i16q, %i17q

which might be seen in a function prolog, or any other place where
the target-specific code generator has preassigned a node. if the
allocator has the misfortune of assigning %rdi to $I17, then a false
edge appears in the interference graph and it will not coalesce them,
leaving at best an extra copy and at worst two useless copies around.

we extend the nop() pass to detect these strays and eliminate them
after the fact. there is probably a better solution within the graph
allocator itself- this fix still leaves an extra node in the graph
during allocation which increases the register pressure- but for the
moment this does the job. the allocator will be getting an overhaul
someday (its spill logic, in particular); we can address this then.

Signed-off-by: Charles E. Youse <charles@gnuless.org>