cc1: kill.c: add READ set to KILL set in kill.c
the KILL() set is the set of registers defined in a block. add
its complement, READ(), as the set of registers used in a block.

create a struct kill to hold the kill/read data in a block and
add the plumbing to init/clear it the same as other block data.

extend the kill_gather() function to allow gathering READs as
well as KILLs if the caller wishes.

Signed-off-by: Charles E. Youse <charles@gnuless.org>