cc1: add preliminary loop-invariant code motion
at the moment, only runs on the machine-independent IR, and is
purposely hobbled to not do too much, lest it interfere with the
target instruction selection. we're more interested in identifying
the loop invariants so we can spot induction variables; code motion
will be repeated on the target IR where it will be less restricted.

Signed-off-by: Charles E. Youse <charles@gnuless.org>