cc1: loop.c: expand loop-invariant code motion to memory loads
if there are no memory writes in the body of a loop, then we can
safely move memory loads out of the loop body, too. like the rest
of LICM at the moment, this isn't a huge win, but it's pretty cheap.

while at it, fixed range_by_def() to return 0 if no range exists.
added supporting functions to check if a block or group of blocks
write to memory. added LICM disqualification for volatile insns.

Signed-off-by: Charles E. Youse <charles@gnuless.org>