cc1: expr.c: do not strip qualifiers when promoting aggregates
per ANSI 6.2.2.1, if an "lvalue has qualified type, the value has
the unqualified version of the type of the lvalue", but this does
NOT apply to aggregates on the LHS of the . operator. promote()
failed to take this into account, not only goofing the semantics,
but crashing the compiler when it tried to cast an aggregate.

Signed-off-by: Charles E. Youse <charles@gnuless.org>