cc1: add support for ALGOL-like statement expressions (a la GCC)
not too many changes required to implement this, though some care
needed to be taken to ensure that such expressions only appear in
function contexts and not in static data initialization. init_local()
needed to be made re-entrant, since an automatic initializer can be a
statement expression, which might itself involve automatic initializers.

updated the README.md to indicate that this is now implemented. took
the opportunity to clarify the text in the vicinity with respect to
planned features.

fixes: #9

Signed-off-by: Charles E. Youse <charles@gnuless.org>