cc1: add support for inline assembly
this commit is probably larger than it should be, but the changes
outside of insn.c aren't intrusive, so i won't bother breaking it up.

from cc1/asm.h:

the syntax of an __asm() instruction is approximately:

'__asm' [ 'volatile' ] '(' string-literal [':' inputs [':' outputs]] ')'

inputs and outputs take the form     ( reg [ '=' variable ] ) | flag

where regs are physical registers, variables are scalars (which must be
class-appropriate to the reg) and flag is one of 'mem' or 'cc'. inputs
indicate USEs, and for reg/variable pairs the compiler arranges for the
variables' values to be placed in reg before the user asm is emitted. the
outputs indicate DEFs, and for those reg/variable pairs, the compiler will
retrieve the updated variables' values from their respective registers,
after the user's assembly. the flag 'cc', and regs without associated
variables, are only allowed as outputs- the meaning is "contents trashed".

if __asm() is volatile, then the I_ASM will be marked INSN_FLAG_VOLATILE
and so the instruction will not be moved/reordered, etc.

fixes: #8

Signed-off-by: Charles E. Youse <charles@gnuless.org>